rag-system/pyproject.toml
[project]
name = "rag-system"
version = "0.1.0"
description = ""
authors = [
    {name = "Julien",email = "juliengmanana@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "llama-index (>=0.14.21,<0.15.0)",
    "pinecone-client (>=6.0.0,<7.0.0)",
    "cohere (>=6.1.0,<7.0.0)",
    "sentence-transformers (>=5.4.1,<6.0.0)",
    "fastapi (>=0.136.1,<0.137.0)",
    "uvicorn (>=0.46.0,<0.47.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "rank-bm25 (>=0.2.2,<0.3.0)",
    "nltk (>=3.9.4,<4.0.0)",
    "psycopg2-binary (>=2.9.12,<3.0.0)",
    "pgvector (>=0.4.2,<0.5.0)",
    "openai (>=2.33.0,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)"
]

[tool.poetry]
packages = [{include = "rag_system", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
